(WIP) simplify jenkins-sandbox
This commit removes the need for a local python venv, docker swarm, and
credentials for the Jenkins instance. Jenkins is deployed without a
user making it simpler to deploy the jobs via jenkins-jobs.

The deployment process is now a build, run, exec sequence that can be
managed by the end user.